requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
python-telegram-bot>=21.0.0
requests>=2.0.0
reportlab==4.4.10
snscrape==0.7.0.20230622
